Rename thread_pool
